refactor: add RepeatMode type
