+++ packages/client/services/converter/index.ts
@@ -2,6 +2,6 @@ export * from './convertArtist';
 export * from './convertEpisode';
 export * from './convertRelease';
 export * from './convertText';
-export * from './convertToContentListItem';
+export * from './convertItem';
 export * from './convertTrack';
 export * from './convertUri';
