+++ packages/client/store/getters.ts
@@ -24,7 +24,7 @@ const getters: VuexGetters<State, Getters> = {
   headerStyles(state) {
     const opacity = 0.6;
     const rgbList = state.dominantBackgroundColor?.rgb
-      ?.map((color) => color * DARKEN_FILTER_RATIO) as App.DominantColor['rgb'] ?? BACKGROUND_COLOR_RGB;
+      ?.map((color) => color * DARKEN_FILTER_RATIO) ?? BACKGROUND_COLOR_RGB as App.DominantColor['rgb'];
     return { backgroundColor: `rgba(${rgbList.join(',')},${opacity})` };
   },
 };
